chore(examples): remove `create-vite-old` example and its related files
