--- examples/create-vite-old/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <!--suppress HtmlUnknownTarget -->
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
-</html>
