--- examples/create-vite-old/src/refreshComponent.tsx
@@ -1,11 +1,0 @@
-const RefreshComponent = () => {
-	console.log("Initial refresh called");
-
-	return (
-		<div>
-			Initial refreshing................
-		</div>
-	)
-}
-
-export default RefreshComponent;
