+++ examples/create-vite/package.json
@@ -15,7 +15,7 @@
     "react-router-dom": "^7.7.1"
   },
   "devDependencies": {
-    "@auth-kit/react-router": "4.0.2-alpha.0",
+    "@auth-kit/react-router": "4.0.2-alpha.1",
     "@eslint/js": "^9.30.1",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
