+++ examples/create-vite/src/componants/Login.jsx
@@ -17,20 +17,20 @@ const Login = () => {
    * This function demonstrates a fake authentication, using useSignIn function
    */
   const loginHandler = () => {
-    // Assuming that, all network Request is successfully, and the user is authenticated
+    // Assuming that, all network Request is successful, and the user is authenticated
 
     signIn({
       auth: {
         token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiZXhwIjoxNzE0NDYyMTczfQ.J8A_jlXgeZQWm_kMP5sqq6tK0nLLWJGpGXVRtm2DNc8'
       },
-      userState: {name: 'Manas Baroi', uid: 123456},
+      userState: {name: 'React Auth Kit', uid: 123456},
       refresh: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiZXhwIjo4MDA4NjA1MTk1fQ.ijw603AjpAqNwnUXmv6YB5L6m5aL-llIgBsTJo-k2r8',
       navigateTo: '/secure'
     })
   }
   // console.log(isAuthenticated())
   if (isAuthenticated()) {
-    // If an authenticated user, then redirect to secure dashboard
+    // If an authenticated user, then redirect to a secure dashboard
 
     return (
       <Navigate to={'/secure'} replace/>
@@ -40,7 +40,7 @@ const Login = () => {
     // For Demonstration; I'm using just a button to log in.
     // In reality, there should be a form, validation, network request and other things
     return (
-      <button className='p-2 border rounded-md bg-cyan-500 hover:bg-cyan-700 text-white' onClick={loginHandler}>Log In!!</button>
+      <button onClick={loginHandler}>Log In!!</button>
     )
   }
 }
